fix(libslirp): Ensure vmstate_size returns a signed integer value
Correctly cast field->size to a signed integer to prevent potential overflow issues when calculating the size in vmstate_size.

+++ src/core_melonds/src/net/libslirp/src/vmstate.c
@@ -282,7 +282,7 @@ static int vmstate_n_elems(char *opaque, const VMStateField *field)
 
 static int vmstate_size(char *opaque, const VMStateField *field)
 {
-    int size = field->size;
+    int size = (int)field->size;
 
     if (field->flags & VMS_VBUFFER) {
         size = *(int32_t *)(opaque + field->size_offset);
